Okay moved menu() and methods menu was calling to ConsoleUtils class. Changed them methods back to private (where before I had changed to protected when they were separated from menu() ).
