+++ CryptoProject/src/main/java/Client/Wallet.java
@@ -111,7 +111,7 @@ public class Wallet {
     }
 
 
-    //TODO Show important wallet data.
+    //TODO (Review)Show important wallet data.
     public void showWalletData() {
 
     }
