+++ CryptoProject/src/main/java/Server/FileOperations.java
@@ -10,7 +10,7 @@ import Client.Wallet;
 
 public class FileOperations {
 
-    //TODO Serialize the wallet out to a file
+    //TODO -Amee-Serialize the wallet out to a file
     private Wallet saveData;
 
     //this method will save the wallet passed into it as a serialized object
@@ -20,7 +20,7 @@ public class FileOperations {
 
     }
 
-    //TODO Deserialize a wallet from file into an object.
+    //TODO -Amee-Deserialize a wallet from file into an object.
     public Wallet loadWallet() {
